fix: small naming issues and typos
