+++ pages/api/exportStats.ts
@@ -1,13 +1,12 @@
 import { withApiAuth } from "../../lib/auth/withApiAuth";
 import { withApiMethods } from "../../lib/withApiMethods";
-import { prisma } from "../../lib/db";
-import { createExcel } from "../../lib/exportStats/createExcel";
+import { convertStatsIntoExcel } from "../../lib/exportStats/convertStatsIntoExcel";
 
 export default withApiAuth(
   withApiMethods({
     GET: async (_req, res) => {
       try {
-        const workbook = await createExcel();
+        const workbook = await convertStatsIntoExcel();
         workbook.write("zsk-amnesty.xlsx", res);
       } catch {
         res.status(500).json({ message: "Internal server error" });
